Establish a visual hierarchy
So far, we have done these sections: Summary, Education, and Technologies.

These sections still need to be formatted for both mobile and desktop sizes: Tools, Achievements, Projects, Contact Me and Socials.

We will set the flex-direction for the Technologies table to column on mobile devices, and row on desktop computers.